+++ plugins/CoreAdminHome/tests/UI/TrackingFailures_spec.js
@@ -8,6 +8,8 @@
 describe("TrackingFailures", function () {
     this.timeout(0);
 
+    this.fixture = 'Piwik\\Tests\\Fixtures\\InvalidVisits';
+
     var manageUrl = '?module=CoreAdminHome&action=trackingFailures&idSite=1&period=day&date=today';
     var widgetUrl = '?module=Widgetize&action=iframe&moduleToWidgetize=CoreAdminHome&actionToWidgetize=getTrackingFailures&idSite=1&period=day&date=today&widget=1';