+++ plugins/CoreHome/javascripts/dataTable.js
@@ -630,11 +630,17 @@ $.extend(DataTable.prototype, UIControl.prototype, {
         var tableRowLimits = piwik.config.datatable_row_limits,
         evolutionLimits =
         {
-            day: [8, 30, 60, 90, 180, 365, 500],
-            week: [4, 12, 26, 52, 104, 500],
+            day: [8, 30, 60, 90, 180],
+            week: [4, 12, 26, 52, 104],
             month: [3, 6, 12, 24, 36, 120],
             year: [3, 5, 10]
         };
+
+        // only allow big evolution limits for non flattened reports
+        if (!parseInt(this.param.flat)) {
+            evolutionLimits.day.push(365, 500);
+            evolutionLimits.week.push(500);
+        }
 
         var self = this;
         if (typeof self.parentId != "undefined" && self.parentId != '') {
@@ -1293,6 +1299,13 @@ $.extend(DataTable.prototype, UIControl.prototype, {
                 setText(this, 'flat', 'CoreHome_UnFlattenDataTable', 'CoreHome_FlattenDataTable');
             })
             .click(generateClickCallback('flat'));
+
+        // handle flatten
+        $('.dataTableShowTotalsRow', domElem)
+            .each(function () {
+                setText(this, 'keep_totals_row', 'CoreHome_RemoveTotalsRowDataTable', 'CoreHome_AddTotalsRowDataTable');
+            })
+            .click(generateClickCallback('keep_totals_row'));
 
         $('.dataTableIncludeAggregateRows', domElem)
             .each(function () {
@@ -1975,6 +1988,7 @@ DataTable.registerFooterIconHandler('ecommerceAbandonedCart', switchToEcommerceV
 DataTable.switchToGraph = function (dataTable, viewDataTable) {
     var filters = dataTable.resetAllFilters();
     dataTable.param.flat = filters.flat;
+    dataTable.param.keep_totals_row = filters.keep_totals_row;
     dataTable.param.columns = filters.columns;
 
     dataTable.param.viewDataTable = viewDataTable;